+++ test/zaikio/procurement_supplier_test.rb
@@ -126,34 +126,6 @@ class Zaikio::ProcurementSupplierTest < ActiveSupport::TestCase
     end
   end
 
-  test "fetching and updating availability checks" do
-    VCR.use_cassette("supplier_material_availabilitc_checks") do
-      Zaikio::Procurement.with_token(valid_token) do
-        availability_check = Zaikio::Procurement::MaterialAvailabilityCheck
-                             .find("40009396-d4e8-48eb-801e-c43f868748e1")
-        assert_equal 500, availability_check.amount
-        assert_equal "pallet", availability_check.unit
-        assert_nil availability_check.responded_at
-
-        Zaikio::Procurement.with_token(valid_token) do
-          availability_check.update(
-            earliest_delivery_date: 20.days.from_now,
-            stock_availability_amount: 1_000,
-            confirmed_price: "89000.40",
-            expires_at: 1.hour.from_now
-          )
-        end
-
-        availability_check = Zaikio::Procurement::MaterialAvailabilityCheck
-                             .find("40009396-d4e8-48eb-801e-c43f868748e1")
-        assert_equal 1_000, availability_check.stock_availability_amount
-        assert_not_nil availability_check.responded_at
-      end
-
-      assert_nil Zaikio::Procurement::AuthorizationMiddleware.token
-    end
-  end
-
   test "fetching and updating order conditions inquiries" do
     VCR.use_cassette("supplier_order_conditions_inquiries") do
       id = "1cb57cfb-4119-4904-a000-3c390c976cd8"