+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- **Breaking change for Supplier API**: Removed MaterialAvailabilityCheck for suppliers
+
 ## [0.4.0]
 
 ### Added